+++ test/net/connect.cpp
@@ -4,7 +4,9 @@
 
 #define CATCH_CONFIG_MAIN
 
-#include "src/net/connect.hpp"
+#include "src/common/check_connectivity.hpp"
+#include "src/net/connect_impl.hpp"
+#include "src/net/emitter.hpp"
 #include "src/ext/Catch/single_include/catch.hpp"
 #include <event2/bufferevent.h>
 #include <iostream>
@@ -78,6 +80,38 @@ TEST_CASE("connect_base deals with bufferevent_socket_connect error") {
             },
             3.14);
     });
+}
+
+static void success(std::string, int, Callback<Error, Var<Transport>> cb,
+                    Settings, Var<Logger> logger, Var<Reactor>) {
+    cb(NoError(), Var<Transport>(new Emitter(logger)));
+}
+
+TEST_CASE("net::connect_many() correctly handles net::connect() success") {
+    Var<ConnectManyCtx> ctx =
+        connect_many_make("www.google.com", 80, 3,
+                          [](Error err, std::vector<Var<Transport>> conns) {
+                              REQUIRE(!err);
+                              REQUIRE(conns.size() == 3);
+                          },
+                          {}, Logger::global(), Reactor::global());
+    connect_many_impl<success>(ctx);
+}
+
+static void fail(std::string, int, Callback<Error, Var<Transport>> cb, Settings,
+                 Var<Logger>, Var<Reactor>) {
+    cb(GenericError(), Var<Transport>(nullptr));
+}
+
+TEST_CASE("net::connect_many() correctly handles net::connect() failure") {
+    Var<ConnectManyCtx> ctx =
+        connect_many_make("www.google.com", 80, 3,
+                          [](Error err, std::vector<Var<Transport>> conns) {
+                              REQUIRE(err);
+                              REQUIRE(conns.size() == 0);
+                          },
+                          {}, Logger::global(), Reactor::global());
+    connect_many_impl<fail>(ctx);
 }
 
 /*
@@ -283,5 +317,157 @@ TEST_CASE("connect() fails when setting an invalid dns") {
                       {{"dns/nameserver", "8.8.8.1"},
                        {"dns/timeout", 0.001},
                        {"dns/attempts", 1}});
+    });
+}
+
+TEST_CASE("net::connect_many() works as expected") {
+    loop_with_initial_event_and_connectivity([]() {
+        connect_many("www.google.com", 80, 3,
+                     [](Error error, std::vector<Var<Transport>> conns) {
+                         REQUIRE(!error);
+                         REQUIRE(conns.size() == 3);
+                         Var<int> n(new int(0));
+                         for (auto conn : conns) {
+                             REQUIRE(static_cast<bool>(conn));
+                             conn->close([n]() {
+                                 if (++(*n) >= 3) {
+                                     break_loop();
+                                 }
+                             });
+                         }
+                     });
+    });
+}
+
+TEST_CASE("net::connect() works with a custom reactor") {
+    // Note: this is how Portolan uses measurement-kit
+    if (!CheckConnectivity::is_down()) {
+        set_verbosity(MK_LOG_DEBUG);
+        Var<Reactor> reactor = Reactor::make();
+        auto ok = false;
+        connect("nexa.polito.it", 22,
+                [&](Error error, Var<Transport> txp) {
+                    if (error) {
+                        reactor->break_loop();
+                        return;
+                    }
+                    txp->close([&]() { reactor->break_loop(); });
+                    ok = true;
+                },
+                {}, Logger::global(), reactor);
+        reactor->loop();
+        REQUIRE(ok);
+    }
+}
+
+TEST_CASE("net::connect() can connect to open port") {
+    loop_with_initial_event_and_connectivity([]() {
+        connect("www.kernel.org", 80, [](Error error, Var<Transport> txp) {
+            REQUIRE(!error);
+            Var<ConnectResult> cr = error.context.as<ConnectResult>();
+            REQUIRE(cr);
+            REQUIRE(!cr->resolve_result.inet_pton_ipv4);
+            REQUIRE(!cr->resolve_result.inet_pton_ipv6);
+            REQUIRE(cr->resolve_result.addresses.size() > 0);
+            REQUIRE(cr->connected_bev);
+            txp->close([]() { break_loop(); });
+        });
+    });
+}
+
+TEST_CASE("net::connect() can connect to ssl port") {
+    loop_with_initial_event_and_connectivity([]() {
+        set_verbosity(MK_LOG_DEBUG);
+        connect("nexa.polito.it", 443,
+                [](Error error, Var<Transport> txp) {
+                    REQUIRE(!error);
+                    Var<ConnectResult> cr = error.context.as<ConnectResult>();
+                    REQUIRE(cr);
+                    REQUIRE(!cr->resolve_result.inet_pton_ipv4);
+                    REQUIRE(!cr->resolve_result.inet_pton_ipv6);
+                    REQUIRE(cr->resolve_result.addresses.size() > 0);
+                    REQUIRE(cr->connected_bev);
+                    txp->close([]() { break_loop(); });
+                },
+                {{"net/ssl", true},
+                 {"net/ca_bundle_path", "test/fixtures/certs.pem"}});
+    });
+}
+
+TEST_CASE("net::connect() ssl fails when presented an expired certificate") {
+    loop_with_initial_event_and_connectivity([]() {
+        set_verbosity(MK_LOG_DEBUG);
+        connect("expired.badssl.com", 443,
+                [](Error error, Var<Transport> ) {
+                    REQUIRE(error);
+                    break_loop();
+                },
+                {{"net/ssl", true},
+                 {"net/ca_bundle_path", "test/fixtures/certs.pem"}});
+    });
+}
+
+TEST_CASE("net::connect() ssl fails when presented a certificate with the "
+          "wrong hostname") {
+    loop_with_initial_event_and_connectivity([]() {
+        set_verbosity(MK_LOG_DEBUG);
+        connect("wrong.host.badssl.com", 443,
+                [](Error error, Var<Transport>) {
+                    REQUIRE(error);
+                    break_loop();
+                },
+                {{"net/ssl", true},
+                 {"net/ca_bundle_path", "test/fixtures/certs.pem"}});
+    });
+}
+
+TEST_CASE("net::connect() ssl works when using SNI") {
+    loop_with_initial_event_and_connectivity([]() {
+        set_verbosity(MK_LOG_DEBUG);
+        connect("sha256.badssl.com", 443,
+                [](Error error, Var<Transport> txp) {
+                    REQUIRE(!error);
+                    txp->close([]() { break_loop(); });
+                },
+                {{"net/ssl", true},
+                 {"net/ca_bundle_path", "test/fixtures/certs.pem"}});
+    });
+}
+
+TEST_CASE("net::connect() ssl works when setting ca_bundle_path via global "
+          "settings") {
+    set_verbosity(MK_LOG_DEBUG);
+    Var<Settings> global_settings = Settings::global();
+    (*global_settings)["net/ca_bundle_path"] = "test/fixtures/certs.pem";
+    loop_with_initial_event_and_connectivity([]() {
+        connect("nexa.polito.it", 443,
+                [](Error error, Var<Transport> txp) {
+                    REQUIRE(!error);
+                    Var<ConnectResult> cr = error.context.as<ConnectResult>();
+                    REQUIRE(cr);
+                    REQUIRE(!cr->resolve_result.inet_pton_ipv4);
+                    REQUIRE(!cr->resolve_result.inet_pton_ipv6);
+                    REQUIRE(cr->resolve_result.addresses.size() > 0);
+                    REQUIRE(cr->connected_bev);
+                    txp->close([]() { break_loop(); });
+                },
+                {{"net/ssl", true}});
+    });
+}
+
+TEST_CASE("net::connect() works in case of error") {
+    loop_with_initial_event_and_connectivity([]() {
+        connect("nexa.polito.it", 81,
+                [](Error error, Var<Transport>) {
+                    REQUIRE(error);
+                    Var<ConnectResult> cr = error.context.as<ConnectResult>();
+                    REQUIRE(cr);
+                    REQUIRE(!cr->resolve_result.inet_pton_ipv4);
+                    REQUIRE(!cr->resolve_result.inet_pton_ipv6);
+                    REQUIRE(cr->resolve_result.addresses.size() > 0);
+                    REQUIRE(!cr->connected_bev);
+                    break_loop();
+                },
+                {{"net/timeout", 5.0}});
     });
 }